*: move pkg/verbose to pkg/logger/log
